Refactor Utils
Created a utils directory and created modules for accumulators and operators
